+++ src/transport/xmpp/xmpp_client.erl
@@ -10,7 +10,7 @@
 -include("xmpp.hrl").
 -include_lib("xmerl/include/xmerl.hrl").
 
--export([start_link/8]).
+-export([start_link/9]).
  
 %% gen_server callbacks
 -export([init/1,
@@ -41,13 +41,15 @@
         % Xmpp server port
         port = 5222,
         % socket mode
-        socket_mod = null
+        socket_mod = null,
+        % reconnect timeout
+        reconnect_timeout = 0
     }).
 
-start_link(CallbackModule, Login, Password, Server, Port, Room, Resource, SocketMode) ->
-    gen_server:start_link(?MODULE, [CallbackModule, Login, Password, Server, Port, Room, Resource, SocketMode], []).
+start_link(CallbackModule, Login, Password, Server, Port, Room, Resource, SocketMode, ReconnectTimeout ) ->
+    gen_server:start_link(?MODULE, [CallbackModule, Login, Password, Server, Port, Room, Resource, SocketMode, ReconnectTimeout], []).
 
-init([CallbackModule, Login, Password, Server, Port, Room, Resource, SocketMode]) ->
+init([CallbackModule, Login, Password, Server, Port, Room, Resource, SocketMode, ReconnectTimeout ]) ->
     % try to connect
     gen_server:cast(self(), {connect, Server, Port}),
     % init process internal state
@@ -58,7 +60,8 @@ init([CallbackModule, Login, Password, Server, Port, Room, Resource, SocketMode]
                 room = list_to_binary(binary_to_list(Room) ++ "/" ++ binary_to_list(Login)),
                 resource = Resource,
                 port = Port,
-                socket_mod = SocketMode
+                socket_mod = SocketMode,
+                reconnect_timeout = ReconnectTimeout
                }
     }.
 
@@ -90,7 +93,8 @@ handle_cast({connect, Host, Port}, State) ->
         {error, Reason} ->
             % Some log
             lager:error("Unable to connect to xmpp server with reason ~p", [Reason]),
-            {noreply, State}
+            % try to reconnect
+            try_reconnect(State)
     end;
 
 %% @doc send message to jabber
@@ -114,6 +118,30 @@ handle_cast({send_message, From, Message}, State) ->
 handle_cast(_Msg, State) ->
     {noreply, State}.
 
+handle_info({ssl_closed, Reason}, State) ->
+    % Some log
+    lager:info("ssl_closed with reason: ~p~n", [Reason]),
+    % try reconnect
+    try_reconnect(State);
+
+handle_info({ssl_error, _Socket, Reason}, State) ->
+    % Some log
+    lager:error("tcp_error: ~p~n", [Reason]),
+    % try reconnect
+    try_reconnect(State);
+
+handle_info({tcp_closed, Reason}, State) ->
+    % Some log
+    lager:info("tcp_closed with reason: ~p~n", [Reason]),
+    % try reconnect
+    try_reconnect(State);
+
+handle_info({tcp_error, _Socket, Reason}, State) ->
+    % Some log
+    lager:error("tcp_error: ~p~n", [Reason]),
+    % try reconnect
+    try_reconnect(State);
+    
 %% @doc Incoming message
 handle_info({_, _Socket, Data}, State) ->
     try
@@ -206,4 +234,19 @@ send_presence(Xml, Socket, M) ->
         _ ->
             pass
     end.
-    
+
+%% @doc try reconnect
+-spec try_reconnect(State :: #state{}) -> {normal, stop, State} | {noreply, State}.
+try_reconnect(#state{reconnect_timeout = Timeout, host = Host, port = Port} = State) ->
+    case Timeout > 0 of
+        true ->
+            % no need in reconnect
+            {normal, stop, State};
+        false ->
+            % sleep
+            timer:sleep(Timeout),
+            % Try reconnect
+            gen_server:cast(self(), {connect, Host, Port}),
+            % return
+            {noreply, State}
+    end.